fix(nx-generators): align and refactor code
